Screen class update

+++ src/ca/javajesus/game/entities/Mob.java
@@ -373,8 +373,8 @@ public class Mob extends Entity {
 				this.y - (int) getBounds().getHeight() / 2 - 2);
 
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2.0;
-		double yOffset = y - modifier / 2.0 - 4;
+		int xOffset = x - modifier / 2;
+		int yOffset = y - modifier / 2 - 4;
 
 		if (isSwimming) {
 			if (isOnFire()) {
