+++ src/ca/javajesus/game/entities/monsters/Centaur.java
@@ -137,8 +137,8 @@ public class Centaur extends Monster {
 		}
 
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2;
-		double yOffset = (y - modifier / 2 - 4) - modifier;
+		int xOffset = x - modifier / 2;
+		int yOffset = (y - modifier / 2 - 4) - modifier;
 
 		if (isDead) {
 			if (isLongitudinal(getDirection())) {
