+++ src/ca/javajesus/game/entities/monsters/Cyclops.java
@@ -106,8 +106,8 @@ public class Cyclops extends Monster {
 		}
 
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2;
-		double yOffset = (y - modifier / 2 - 4) - modifier;
+		int xOffset = x - modifier / 2;
+		int yOffset = (y - modifier / 2 - 4) - modifier;
 
 		int yTile = this.yTile;
 		if (isDead) {
