+++ src/ca/javajesus/game/entities/monsters/Demon.java
@@ -134,8 +134,8 @@ public class Demon extends Monster {
 		}
 
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2;
-		double yOffset = (y - modifier / 2 - 4) - modifier;
+		int xOffset = x - modifier / 2;
+		int yOffset = (y - modifier / 2 - 4) - modifier;
 
 		if (isShooting)
 			xTile += 12;
