+++ src/ca/javajesus/game/entities/monsters/GangMember.java
@@ -139,8 +139,8 @@ public class GangMember extends Monster {
 		}
 
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2;
-		double yOffset = (y - modifier / 2 - 4) - modifier;
+		int xOffset = x - modifier / 2;
+		int yOffset = (y - modifier / 2 - 4) - modifier;
 
 		if (isDead) {
 			isShooting = false;
