+++ src/ca/javajesus/game/entities/projectiles/Missile.java
@@ -14,17 +14,17 @@ public class Missile extends Projectile {
 
 	public Missile(Level level, double x, double y, double xPos, double yPos,
 			Mob mob, double damage) {
-		super(level, 6, 6, 1, Colors.get(-1, Colors.fromHex("#000000"), 
-				Colors.fromHex("#5b5b5b"), Colors.fromHex("#ffea02")), x, y, 5, xPos, yPos,
-				mob, damage);
+		super(level, 6, 6, 1, Colors.get(-1, Colors.fromHex("#000000"),
+				Colors.fromHex("#5b5b5b"), Colors.fromHex("#ffea02")), x, y, 5,
+				xPos, yPos, mob, damage);
 		sound.fire(sound.gunshot);
 	}
 
 	public Missile(Level level, double x, double y, int direction, Mob mob,
 			double damage) {
-		super(level, 6, 6, 1, Colors.get(-1, Colors.fromHex("#ffea02"), 
-				Colors.fromHex("#5b5b5b"), Colors.fromHex("#000000")), x, y, 5, direction,
-				mob, damage);
+		super(level, 6, 6, 1, Colors.get(-1, Colors.fromHex("#ffea02"),
+				Colors.fromHex("#5b5b5b"), Colors.fromHex("#000000")), x, y, 5,
+				direction, mob, damage);
 		sound.fire(sound.gunshot);
 		switch (direction) {
 		case 0:
@@ -41,9 +41,9 @@ public class Missile extends Projectile {
 			return;
 		}
 	}
-	
-public void render(Screen screen) {
-		
+
+	public void render(Screen screen) {
+
 		renderOnTop = true;
 
 		if (hasCollided((int) x, (int) y)) {
@@ -88,7 +88,7 @@ public void render(Screen screen) {
 			}
 
 		}
-		screen.render(this.x, this.y, tileNumber, color, 1, 1,
+		screen.render((int) this.x, (int) this.y, tileNumber, color, 1, 1,
 				sheet);
 	}
