+++ src/ca/javajesus/game/entities/vehicles/Boat.java
@@ -87,8 +87,8 @@ public class Boat extends Vehicle {
 		int flipTop = 0;
 		int flipBottom = 0;
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2.0;
-		double yOffset = y - modifier / 2.0 - 4;
+		int xOffset = x - modifier / 2;
+		int yOffset = y - modifier / 2 - 4;
 
 		if (getDirection() == Direction.NORTH) {
 			xTile += 14;
