+++ src/ca/javajesus/game/entities/vehicles/CenturyLeSabre.java
@@ -95,8 +95,8 @@ public class CenturyLeSabre extends Vehicle {
 		int flipTop = 0;
 		int flipBottom = 0;
 		int modifier = 8 * scale;
-		double xOffset = x - modifier / 2.0;
-		double yOffset = y - modifier / 2.0 - 4;
+		int xOffset = x - modifier / 2;
+		int yOffset = y - modifier / 2 - 4;
 
 		if (getDirection() == Direction.NORTH) {
 			xTile += 14;
