+++ src/ca/javajesus/items/Bazooka.java
@@ -30,8 +30,8 @@ public class Bazooka extends Gun{
 		int flipTop = (player.getNumSteps() >> walkingAnimationSpeed) & 1;
 		int flipBottom = (player.getNumSteps() >> walkingAnimationSpeed) & 1;
 		int modifier = 8 * player.getScale();
-		double xOffset = player.getX() - modifier / 2.0;
-		double yOffset = player.getY() - modifier / 2.0 - 4;
+		int xOffset = player.getX() - modifier / 2;
+		int yOffset = player.getY() - modifier / 2 - 4;
 		int xx = 0;
 
 		if (player.shootingDir == 0) {
